+++ SBO-Lib/lib/SBO/Lib/Readme.pm
@@ -235,11 +235,7 @@ sub get_user_group {
         push @parse_script, "echo $escaped_cmd";
       }
       my ($fh, $parse_script);
-      if ($< == 0) {
-        ($fh, $parse_script) = tempfile(DIR => $tempdir);
-      } else {
-        ($fh, $parse_script) = tempfile(DIR => "/tmp");
-      }
+      ($fh, $parse_script) = tempfile(DIR => "/tmp");
       my $exit;
       ($fh, $exit) = open_fh($parse_script, '>');
       error_code("SlackBuild had variable UID/GID specifications, and writing a parse script failed.", _ERR_OPENFH) if $exit;